Refactor : Separate Package Admin and Branch
