+++ src/main/java/com/erp/process/branch/AuthProcess.java
@@ -1,4 +1,4 @@
-package com.erp.process;
+package com.erp.process.branch;
 
 import org.springframework.http.ResponseEntity;
