+++ src/main/java/com/erp/process/branch/AuthProcessImpl.java
@@ -1,4 +1,4 @@
-package com.erp.process;
+package com.erp.process.branch;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
