+++ src/main/java/com/erp/process/branch/DashboardProcess.java
@@ -1,4 +1,4 @@
-package com.erp.process;
+package com.erp.process.branch;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
