+++ src/main/java/com/erp/process/branch/MemberProcess.java
@@ -1,4 +1,4 @@
-package com.erp.process;
+package com.erp.process.branch;
 
 import java.util.HashMap;
 import java.util.List;
