+++ src/main/java/com/erp/process/branch/ServiceProcess.java
@@ -1,4 +1,4 @@
-package com.erp.process;
+package com.erp.process.branch;
 
 import java.util.List;
 import java.util.stream.Collectors;
